fix overflow with fixed sized integers
calculating mid by summing l and r and dividing by 2 can cause an
error with fixed sized integers if l and r are large enough for
the addition to overflow. this can be worked around by using the
formula l + (r - l) / 2 to calculate mid instead. this overflow
behaviour can be modeled in z3 using BitVec for the two pointers
instead of Int with is an arbitrary size integer.